Conflicts:
	app/Http/Controllers/LocatoreController.php
	database/migrations/2022_05_19_174349_offers_migration.php
	database/seeds/DatabaseSeeder.php
	public/assets/css/style.css
	resources/views/catalog.blade.php
	resources/views/offer/insert.blade.php
	routes/web.php

+++ app/Http/Controllers/LocatarioController.php
@@ -31,9 +31,12 @@ class LocatarioController extends Controller {
     }
 
     public function deleteOption($offerId) {
-        
+        $autenticati = Option::where('offerta_id', $offerId)->value('locatario_id');
+        if ( Auth::id() == $autenticati){
         Option::where('offerta_id',$offerId)->delete();
         return redirect()->action('LocatarioController@showOptionedList');
+        }
+        return redirect()->action('UserController@index');
     }
 
 }
