+++ app/Http/Controllers/LocatoreController.php
@@ -6,7 +6,7 @@ use App\Models\OfferList;
 use App\Models\Resources\Offer;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Catalog;
-use App\Http\Requests\OfferRequest;
+use App\Http\Requests\Request;
 
 class LocatoreController extends Controller {
 
@@ -26,19 +26,30 @@ class LocatoreController extends Controller {
                         ->with('idLarioList', $this->offerListModel->getUserByOffer(Auth::id()))
                         ->with('optionList', $this->offerListModel->getAllOption());
     }
-
-    public function deleteOffer($offerId) {
-
-        Offer::where('offerta_id', $offerId)->delete();
+    
+    public function acceptOffer($opt_id) {
+        
+        $this->offerListModel->setAcceptById($opt_id);
         return redirect()->action('LocatoreController@showOfferList');
     }
-    
+
+    public function deleteOffer($offerId) {
+        
+        $autenticati = Offer::where('offerta_id', $offerId)->value('user_id');
+        
+        if ( Auth::id() == $autenticati){
+         Offer::where('offerta_id', $offerId)->delete();
+         return redirect()->action('LocatoreController@showOfferList');
+        }
+        return redirect()->action('UserController@index');
+    }
+
     public function showInsertOfferForm() {
-        return view('offer/insert');
+        return view('form/offerform');
     }
     
     public function showUpdateOfferForm() {
-        return view('offer/insert');
+        return view('form/offerform');
     }
     
     public function insertOffer(OfferRequest $request) {
